refactor: add btp package for btp related logic, use static imports for well-known constants, remove unnecessary dependency version declaration from pom.xml
