+++ src/main/java/org/example/learningsystem/btp/featureflags/dto/FlagDto.java
@@ -1,4 +1,4 @@
-package org.example.learningsystem.core.featureflags.dto;
+package org.example.learningsystem.btp.featureflags.dto;
 
 public record FlagDto(Integer httpStatus, String featureName, String type, String variation) {
 }
