+++ src/main/java/org/example/learningsystem/btp/featureflags/service/FeatureFlagsServiceImpl.java
@@ -1,8 +1,8 @@
-package org.example.learningsystem.core.featureflags.service;
+package org.example.learningsystem.btp.featureflags.service;
 
-import org.example.learningsystem.core.featureflags.exception.FeatureFlagTypeMismatchException;
-import org.example.learningsystem.core.featureflags.config.FeatureFlagsProperties;
-import org.example.learningsystem.core.featureflags.dto.FlagDto;
+import org.example.learningsystem.btp.featureflags.exception.FeatureFlagTypeMismatchException;
+import org.example.learningsystem.btp.featureflags.config.FeatureFlagsProperties;
+import org.example.learningsystem.btp.featureflags.dto.FlagDto;
 import org.springframework.context.annotation.Profile;
 import org.springframework.http.HttpHeaders;
 import org.springframework.retry.annotation.Recover;
